+++ jib-core/src/main/java/com/google/cloud/tools/jib/configuration/BuildConfiguration.java
@@ -66,6 +66,7 @@ public class BuildConfiguration {
     private String toolName = DEFAULT_TOOL_NAME;
     private EventHandlers eventHandlers = EventHandlers.NONE;
     @Nullable private ExecutorService executorService;
+    private boolean alwaysCacheBaseImage = false;
 
     private Builder() {}
 
@@ -173,6 +174,19 @@ public class BuildConfiguration {
     }
 
     /**
+     * Controls the optimization which skips downloading base image layers that exist in a target
+     * registry. If the user does not set this property then read as false.
+     *
+     * @param alwaysCacheBaseImage if {@code true}, base image layers are always pulled and cached.
+     *     If {@code false}, base image layers will not be pulled/cached if they already exist on
+     *     the target registry.
+     * @return this
+     */
+    public Builder setAlwaysCacheBaseImage(boolean alwaysCacheBaseImage) {
+      this.alwaysCacheBaseImage = alwaysCacheBaseImage;
+      return this;
+    }
+    /**
      * Sets the layers to build.
      *
      * @param layerConfigurations the configurations for the layers
@@ -267,7 +281,8 @@ public class BuildConfiguration {
               layerConfigurations,
               toolName,
               eventHandlers,
-              Preconditions.checkNotNull(executorService));
+              Preconditions.checkNotNull(executorService),
+              alwaysCacheBaseImage);
 
         case 1:
           throw new IllegalStateException(missingFields.get(0) + " is required but not set");
@@ -321,6 +336,7 @@ public class BuildConfiguration {
   private final String toolName;
   private final EventHandlers eventHandlers;
   private final ExecutorService executorService;
+  private final boolean alwaysCacheBaseImage;
 
   /** Instantiate with {@link #builder}. */
   private BuildConfiguration(
@@ -336,7 +352,8 @@ public class BuildConfiguration {
       ImmutableList<LayerConfiguration> layerConfigurations,
       String toolName,
       EventHandlers eventHandlers,
-      ExecutorService executorService) {
+      ExecutorService executorService,
+      boolean alwaysCacheBaseImage) {
     this.baseImageConfiguration = baseImageConfiguration;
     this.targetImageConfiguration = targetImageConfiguration;
     this.additionalTargetImageTags = additionalTargetImageTags;
@@ -350,6 +367,7 @@ public class BuildConfiguration {
     this.toolName = toolName;
     this.eventHandlers = eventHandlers;
     this.executorService = executorService;
+    this.alwaysCacheBaseImage = alwaysCacheBaseImage;
   }
 
   public ImageConfiguration getBaseImageConfiguration() {
@@ -423,6 +441,15 @@ public class BuildConfiguration {
    */
   public boolean isOffline() {
     return offline;
+  }
+
+  /**
+   * Gets whether or not to force caching the base images.
+   *
+   * @return {@code true} if the user wants to force the build to always pull the image layers.
+   */
+  public boolean getAlwaysCacheBaseImage() {
+    return alwaysCacheBaseImage;
   }
 
   /**